skeleton loaders for recent chats
